# backend/docs/swagger.yaml
basePath: /
...
host: localhost:8080
info:
  contact:
    email: support@iac-platform.com
    name: API Support
  description: Infrastructure as Code Platform API Documentation
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: IAC Platform API
  version: "1.0"
paths:
...
  /api/v1/workspaces/{id}/tasks:
    get:
      consumes:
      - application/json
      description: 获取工作空间的任务列表，支持分页、搜索和过滤
      parameters:
      - description: 工作空间ID
        in: path
        name: id
        required: true
        type: integer
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 搜索关键词
        in: query
        name: search
        type: string
      - description: 状态过滤
        in: query
        name: status
        type: string
      - description: 任务类型过滤
        in: query
        name: task_type
        type: string
      - description: 开始日期（RFC3339格式）
        in: query
        name: start_date
        type: string
      - description: 结束日期（RFC3339格式）
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回任务列表
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 无效的工作空间ID
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: 获取任务列表
      tags:
      - Workspace Task
  /api/v1/workspaces/{id}/tasks/{task_id}:
    get:
      consumes:
      - application/json
      description: 根据ID获取任务的详细信息
      parameters:
      - description: 工作空间ID
        in: path
        name: id
        required: true
        type: integer
      - description: 任务ID
        in: path
        name: task_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回任务详情
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 无效的参数
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 任务不存在
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: 获取任务详情
      tags:
      - Workspace Task
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"